requirements.txt
flask==3.0.2
ultralytics==8.1.27
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0 